+++ packages/website/src/_services/maps/map-tool-menu/index.ts
@@ -1,0 +1,1 @@
+export { default as MapToolMenuComponent } from './map-tool-menu.vue'