+++ packages/website/src/species-richness/species-richness.ts
@@ -3,6 +3,7 @@ import { Options, Vue } from 'vue-class-component'
 
 import { Site, Species } from '~/api'
 import { getSpeciesRichnessData, SpeciesRichnessData, TimeBucket } from '~/api/species-richness-service'
+import { TAXONOMY_CLASS_ALL, TAXONOMY_CLASSES } from '~/api/taxonomy-service'
 import { GroupedBarChartItem, HorizontalBarChartComponent } from '~/charts/horizontal-bar-chart'
 import { ColoredFilter, Filter } from '~/dataset-filters'
 import { ComparisonListComponent } from '~/dataset-filters/comparison-list'
@@ -46,7 +47,7 @@ export default class SpeciesRichnessPage extends Vue {
     // TODO 117 - Only update the changed dataset
     const datasets = await Promise.all(
       filters.map(async (filter) => {
-        const { startDate, endDate, sites, otherFilters, color } = filter
+        const { startDate, endDate, sites, color } = filter
         const data = await getSpeciesRichnessData(filterToDataset(filter))
         return { startDate, endDate, sites, color, data }
       })
@@ -76,7 +77,20 @@ export default class SpeciesRichnessPage extends Vue {
   }
 
   getMapDataset (datasets: ColoredDataset[]): MapDataSet[] {
-    return datasets.map(({ color, data, ...filter }) => ({ color, data: data.speciesBySite, ...filter }))
+    const intermediate = datasets.map(({ color, data: srData, ...filter }) => {
+      const data = srData.speciesBySite.map(s => ({
+        ...s,
+        distinctSpecies: {
+          ...s.distinctSpecies,
+          [TAXONOMY_CLASS_ALL.name]: Object.values(s.distinctSpecies).reduce((sum, val) => (sum as number) + (val as number), 0)
+        }
+      }))
+      return { color, data, ...filter, maxValues: {} }
+    })
+    // TODO 209 - Do this natively in the API instead of after the fact
+    const allTaxonClassNames = TAXONOMY_CLASSES.map(c => c.name)
+    const maxAll = Math.max(...intermediate.map(ds => Math.max(...ds.data.map(d => d.distinctSpecies[TAXONOMY_CLASS_ALL.name] as number))))
+    return intermediate.map(ds => ({ ...ds, maxValues: Object.fromEntries(allTaxonClassNames.map(n => [n, maxAll])) }))
   }
 
   getTableData (datasets: ColoredDataset[]): DetectedSpeciesItem[] {